safari-600.7-branch/LayoutTests/http/tests/security/xssAuditor/script-tag-safe2-expected.txt

ALERT: /This is a safe script./
